rename and swap folder
