+++ src/Application/gkNextBenchmark/BenchMark.cpp
@@ -1,7 +1,7 @@
 #include "BenchMark.hpp"
 #include "Options.hpp"
 #include "Utilities/Console.hpp"
-#include "Vulkan/VulkanBaseRenderer.hpp"
+#include "Rendering/VulkanBaseRenderer.hpp"
 #include <fmt/format.h>
 #include <fmt/chrono.h>
