+++ src/Editor/EditorInterface.cpp
@@ -32,7 +32,7 @@
 #include "Utilities/Math.hpp"
 #include "Vulkan/ImageView.hpp"
 #include "Vulkan/RenderImage.hpp"
-#include "Vulkan/VulkanBaseRenderer.hpp"
+#include "Rendering/VulkanBaseRenderer.hpp"
 #include "Editor/IconsFontAwesome6.h"
 
 extern std::unique_ptr<Vulkan::VulkanBaseRenderer> GApplication;
