+++ src/Rendering/LegacyDeferred/SoftwareModernRenderer.cpp
