+++ src/Rendering/ModernDeferred/SoftwareTracingRenderer.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "Vulkan/VulkanBaseRenderer.hpp"
-#include "Vulkan/RayTracing/RayTraceBaseRenderer.hpp"
+#include "Rendering/VulkanBaseRenderer.hpp"
+#include "Rendering/RayTraceBaseRenderer.hpp"
 
 #include <vector>
 #include <memory>
