+++ src/Rendering/VulkanBaseRenderer.cpp
@@ -1,19 +1,29 @@
 #include "VulkanBaseRenderer.hpp"
-#include "Buffer.hpp"
-#include "CommandPool.hpp"
-#include "CommandBuffers.hpp"
-#include "DebugUtilsMessenger.hpp"
-#include "DepthBuffer.hpp"
-#include "Device.hpp"
-#include "Fence.hpp"
-#include "FrameBuffer.hpp"
-#include "Instance.hpp"
-#include "PipelineLayout.hpp"
-#include "RenderPass.hpp"
-#include "Semaphore.hpp"
-#include "Surface.hpp"
-#include "SwapChain.hpp"
-#include "Window.hpp"
+#include "Vulkan/Buffer.hpp"
+#include "Vulkan/CommandPool.hpp"
+#include "Vulkan/CommandBuffers.hpp"
+#include "Vulkan/DebugUtilsMessenger.hpp"
+#include "Vulkan/DepthBuffer.hpp"
+#include "Vulkan/Device.hpp"
+#include "Vulkan/Fence.hpp"
+#include "Vulkan/FrameBuffer.hpp"
+#include "Vulkan/Instance.hpp"
+#include "Vulkan/PipelineLayout.hpp"
+#include "Vulkan/RenderPass.hpp"
+#include "Vulkan/Semaphore.hpp"
+#include "Vulkan/Surface.hpp"
+#include "Vulkan/SwapChain.hpp"
+#include "Vulkan/Window.hpp"
+
+#include "Vulkan/DescriptorSetManager.hpp"
+#include "Vulkan/DescriptorSets.hpp"
+#include "Vulkan/Enumerate.hpp"
+#include "Vulkan/ImageMemoryBarrier.hpp"
+#include "Vulkan/RenderImage.hpp"
+#include "Vulkan/SingleTimeCommands.hpp"
+#include "Vulkan/Strings.hpp"
+#include "Vulkan/Version.hpp"
+
 #include "Assets/Model.hpp"
 #include "Assets/Scene.hpp"
 #include "Assets/UniformBuffer.hpp"
@@ -23,21 +33,13 @@
 #include <array>
 #include <fmt/format.h>
 
-#include "DescriptorSetManager.hpp"
-#include "DescriptorSets.hpp"
-#include "Enumerate.hpp"
-#include "ImageMemoryBarrier.hpp"
 #include "Options.hpp"
-#include "RenderImage.hpp"
-#include "SingleTimeCommands.hpp"
-#include "Strings.hpp"
-#include "Version.hpp"
 #include "HybridDeferred/HardwareTracingRenderer.hpp"
 #include "LegacyDeferred/SoftwareModernRenderer.hpp"
 #include "ModernDeferred/SoftwareTracingRenderer.hpp"
 #include "RayQuery/PathTracingRenderer.hpp"
 #include "Runtime/Engine.hpp"
-#include "Vulkan/PipelineCommon/CommonComputePipeline.hpp"
+#include "Rendering/PipelineCommon/CommonComputePipeline.hpp"
 
 namespace
 {
